+++ src/input/ndsi.c
@@ -1,11 +1,11 @@
 
 #include "../unilayer.h"
 
-int input_init() {
+uint8_t input_init() {
    return 1;
 }
 
-int input_poll() {
+uint8_t input_poll() {
    return 0;
 }
